scripts/vertexai/test_vertexai.py
"""
Test script for Vertex AI integration.
"""
from google.cloud import aiplatform
import os

PROJECT_ID = os.getenv("VERTEXAI_PROJECT_ID")
LOCATION = os.getenv("VERTEXAI_LOCATION")
MODEL_NAME = os.getenv("VERTEXAI_MODEL_NAME", "gemma-3-9b-it-e4b")

# TODO: Implement a test call to Vertex AI model
