requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
cachetools==5.3.3
google-cloud-aiplatform==1.44.0
google-cloud-firestore==2.16.0
google-cloud-secret-manager==2.18.4
httpx==0.27.0

# Linting & Formatting
ruff==0.4.2

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

# Type Checking
mypy==1.9.0

# For MCP microservices (example, can be split per service)
